pyproject.toml
[tool.poetry]
name = "OmnyBase autoconfig"
version = "0.1.0"
description = ""
authors = ["Dmitry Davydkin"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
onvif-zeep-async = "^1.2.0"
asyncio = "^3.4.3"
fastapi = "^0.70.0"
pydantic = "^1.8.2"
httpx = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
